added skip ecopcr and blast cli arguments

+++ crux_db/crux.sh
@@ -31,8 +31,10 @@ ID2=""
 COV2=""
 HEADER=""
 HELP=""
+ECOPCRDIR=""
+SKIPBLAST="FALSE"
 
-while getopts "n:f:r:s:m:o:d:q?:u:l?:k:e:g:t:v:i:c:a:z:y:j:w:p:x:b:h?:" opt; do
+while getopts "n:f:r:s:m:o:d:q?:u:l?:k:e:g:t:v:i:c:a:z:y:j:w:p:x:b:h?:1:2" opt; do
     case $opt in
         n) NAME="$OPTARG"
         ;;
@@ -86,6 +88,9 @@ while getopts "n:f:r:s:m:o:d:q?:u:l?:k:e:g:t:v:i:c:a:z:y:j:w:p:x:b:h?:" opt; do
         ;;
         h) HELP="TRUE"
         ;;
+        1) ECOPCRDIR="$OPTARG"
+        ;;
+        2) SKIPBLAST="TRUE"
     esac
 done
 
@@ -160,14 +165,8 @@ else
   exit
 fi
 
-
 source $DB/scripts/crux_vars.sh
 source $DB/scripts/crux_config.sh
-
-if [ ! -d "${OBI_DB}" ]; then
-    >&2 echo "ERROR: directory ${OBI_DB} not valid. Please check your configuration file."
-    exit 1
-fi
 
 ${MODULE_SOURCE}
 ${QIIME}
@@ -189,39 +188,30 @@ echo " "
 echo " "
 echo "Part 1.1:"
 echo "Run ecoPCR with ${NAME} primers F- ${FP} R- ${RP} and these parameters:"
-echo "     mismatch = ${ERROR:=$ECOPCR_e}"
+echo "     missmatch = ${ERROR:=$ECOPCR_e}"
 echo "     expected amplicon length between ${SHRT} and ${LNG}"
 echo ""
 ###
 mkdir -p ${ODIR}/${NAME}_ecoPCR
 mkdir -p ${ODIR}/${NAME}_ecoPCR/raw_out/
-#run ecoPCR on each folder in the obitools database folder
 
-for dn in $(find "${OBI_DB}" -mindepth 1 -maxdepth 1 -type d -name "OB_dat_*"); do
-    dn="${dn##*/}"
-    echo "Processing directory ${dn}..."
-    for fn in $(find "${OBI_DB}/${dn}" -mindepth 1 -maxdepth 1 -type f -name "OB_dat_*.tdx"); do
-        # Trim off file extension
-        fn="${fn%.tdx}"
-        fn="${fn##*/}"
-        echo "Running ecoPCR on file ${fn}..."
-        ${ecoPCR} -d "${OBI_DB}/${dn}/${fn}" -e ${ERROR:=$ECOPCR_e} -l ${SHRT} -L ${LNG} ${FP} ${RP} -D 1 > ${ODIR}/${NAME}_ecoPCR/raw_out/${NAME}_${fn}_ecoPCR_out
-        echo "...${fn} ecoPCR is finished"
-        echo ""
-        date
-    done 
-done
-
-#for db in ${OBI_DB}/OB_dat_*/
-#do
-# db1=${db%/}
-# j=${db1#${OBI_DB}/}
-# echo "..."${j}" ecoPCR is running"
-# ${ecoPCR} -d ${db}${j} -e ${ERROR:=$ECOPCR_e} -l ${SHRT} -L ${LNG} ${FP} ${RP} -D 1 > ${ODIR}/${NAME}_ecoPCR/raw_out/${NAME}_${j}_ecoPCR_out
-# echo "..."${j}" ecoPCR is finished"
-# echo ""
-#date
-#done
+if [ "${ECOPCRDIR}" = "" ]
+then
+  #run ecoPCR on each folder in the obitools database folder
+  for db in ${OBI_DB}/OB_dat_*/
+  do
+  db1=${db%/}
+  j=${db1#${OBI_DB}/}
+  echo "..."${j}" ecoPCR is running"
+  ${ecoPCR} -d ${db}${j} -e ${ERROR:=$ECOPCR_e} -l ${SHRT} -L ${LNG} ${FP} ${RP} -D 1 > ${ODIR}/${NAME}_ecoPCR/raw_out/${NAME}_${j}_ecoPCR_out
+  echo "..."${j}" ecoPCR is finished"
+  echo ""
+  date
+  done
+else
+  echo "Skipping ecoPCR Part 1.1... copying ${ECOPCRDIR} to ${ODIR}/${NAME}_ecoPCR"
+  cp -r ${ECOPCRDIR} ${ODIR}/${NAME}_ecoPCR
+fi
 ###
 
 ##########################
@@ -263,115 +253,120 @@ echo " "
 echo "Part 1.3:"
 echo "Clean ${NAME} ecoPCR output for blasting and run BLAST jobs"
 
-## loop through ecoPCR results
-
-if [ "${LOCALMODE}" = "FALSE" ];
+if [ "${SKIPBLAST}" = "FALSE" ]
 then
- for str in ${ODIR}/${NAME}_ecoPCR/cleaned/*_ecoPCR_blast_input_a_and_g_clean.fasta
- do
-  str1=${str%_ecoPCR_blast_input_a_and_g_clean.fasta}
-  j=${str1#${ODIR}/${NAME}_ecoPCR/cleaned/}
-  mkdir -p ${ODIR}/${NAME}_ecoPCR/cleaned/${j}
+  ## loop through ecoPCR results
 
-  if [ -s "${str}" ];
+  if [ "${LOCALMODE}" = "FALSE" ];
   then
-    mkdir -p ${ODIR}/${NAME}_BLAST/
-    mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out
-    mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out/raw
-    mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out/fasta
-    echo " "
-    echo "${str} has ecoPCR reads that passed the minimum criteria to move to the next step."
-    # split ecopcr output into files with 500 reads each
-    split -l 1000 ${str} ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_
-     # do something as file has data
-    i=1
-    for nam in ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_*
-  	do
-	    nam1=${nam%_*}
-	 # submit blast jobs for each file, and then remove reads with duplicate accession version numbers
-      cp ${nam} ${nam1}_${i}
-      rm ${nam}
-      echo "${nam1}_${i}_blast_1.complete" >> ${ODIR}/${NAME}_BLAST/blast_complete_outfiles.txt
-      echo "${nam1}_${i}_blast_2.complete" >> ${ODIR}/${NAME}_BLAST/blast_complete_outfiles.txt
-      ((i=i+1))
-    done
-    # if local blast in line
-    # if not local submit blast array jobs
-    ### count number of files in the directory
-    file_count=$( shopt -s nullglob ; set -- ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_* ; echo $#)
-    # submit blast jobs for each file, and then remove reads with duplicate accession version numbers
-    array_var="\$SGE_TASK_ID"
-    printf "${BLAST1_HEADER}\n#$ -t 1-${file_count}\n\n\n/bin/bash ${DB}/scripts/sub_blast1.sh -n ${NAME} -q ${nam1}_${array_var} -o ${ODIR} -k ${j} -l blast_ready_${array_var} -d ${DB} -v ${EVAL1:=$BLAST1_eVALUE} -t ${THREAD1:=$BLAST1_NUM_THREADS} -i ${ID1:=$BLAST1_PERC_IDENTITY} -c ${COV1:=$BLAST1_HSP_PERC} -a ${RETURN:=$BLAST1_NUM_ALIGNMENTS} -y ${GO:=$GAP_OPEN} -z ${GE:=$GAP_EXTEND}\n" > ${ODIR}/Run_info/blast_jobs/blast1_${j}.sh
-    ${QUEUESUBMIT} ${ODIR}/Run_info/blast_jobs/blast1_${j}.sh
-    printf "${BLAST2_HEADER}\n#$ -t 1-${file_count}\n\n\n/bin/bash ${DB}/scripts/sub_blast2.sh -n ${NAME} -q ${nam1}_${array_var} -o ${ODIR} -k ${j} -l blast_ready_${array_var} -d ${DB} -w ${EVAL2:=$BLAST2_eVALUE} -j ${THREAD2:=$BLAST2_NUM_THREADS} -p ${ID2:=$BLAST2_PERC_IDENTITY} -x ${COV2:=$BLAST2_HSP_PERC} -a ${RETURN:=$BLAST2_NUM_ALIGNMENTS} -y ${GO:=$GAP_OPEN} -z ${GE:=$GAP_EXTEND}\n" > ${ODIR}/Run_info/blast_jobs/blast2_${j}.sh
-    ${QUEUESUBMIT} ${ODIR}/Run_info/blast_jobs/blast2_${j}.sh
-  else
-    echo " "
-    echo "${str} did not pass the minimum criteria that passes ecoPCR reads to the next step."
-    rm ${str}
-  fi
- done
-else
   for str in ${ODIR}/${NAME}_ecoPCR/cleaned/*_ecoPCR_blast_input_a_and_g_clean.fasta
   do
-   str1=${str%_ecoPCR_blast_input_a_and_g_clean.fasta}
-   j=${str1#${ODIR}/${NAME}_ecoPCR/cleaned/}
-   mkdir -p ${ODIR}/${NAME}_ecoPCR/cleaned/${j}
-   echo ""
-   echo "run BLAST inline"
-   if [ -s "${str}" ];
-   then
-     mkdir -p ${ODIR}/${NAME}_BLAST/
-     mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out
-     mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out/raw
-     mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out/fasta
-     echo " "
-     echo "${str} has ecoPCR reads that passed the minimum criteria."
-     # split ecopcr output into files with 500 reads each
-     split -l 1000 ${str} ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_
-      # do something as file has data
-     i=1
-     for nam in ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_*
-     do
-       nam1=${nam%_*}
-    # submit blast jobs for each file, and then remove reads with duplicate accession version numbers
-       cp ${nam} ${nam1}_${i}
-       rm ${nam}
-       echo ""
-       echo "Running BLAST1 on ${nam1}_${i}"
-       date
-       printf "#!/bin/bash\n\n\n/bin/bash ${DB}/scripts/sub_blast1.sh -n ${NAME} -q ${nam1}_${i} -o ${ODIR} -k ${j} -l blast_ready_${i} -d ${DB} -v ${EVAL1:=$BLAST1_eVALUE} -t ${THREAD1:=$BLAST1_NUM_THREADS} -i ${ID1:=$BLAST1_PERC_IDENTITY} -c ${COV1:=$BLAST1_HSP_PERC} -a ${RETURN:=$BLAST1_NUM_ALIGNMENTS} -y ${GO:=$GAP_OPEN} -z ${GE:=$GAP_EXTEND}\n" > ${ODIR}/Run_info/blast_jobs/blast1_${j}_${i}.sh
-       /bin/bash ${ODIR}/Run_info/blast_jobs/blast1_${j}_${i}.sh
-       echo "Running BLAST2 on ${nam1}_${i}"
-       date
-       printf "#!/bin/bash\n\n\n/bin/bash ${DB}/scripts/sub_blast2.sh -n ${NAME} -q ${nam1}_${i} -o ${ODIR} -k ${j} -l blast_ready_${i} -d ${DB} -w ${EVAL2:=$BLAST2_eVALUE} -j ${THREAD2:=$BLAST2_NUM_THREADS} -p ${ID2:=$BLAST2_PERC_IDENTITY} -x ${COV2:=$BLAST2_HSP_PERC} -a ${RETURN:=$BLAST2_NUM_ALIGNMENTS} -y ${GO:=$GAP_OPEN} -z ${GE:=$GAP_EXTEND}\n" > ${ODIR}/Run_info/blast_jobs/blast2_${j}_${i}.sh
-       /bin/bash ${ODIR}/Run_info/blast_jobs/blast2_${j}_${i}.sh
-       ((i=i+1))
-     done
-     date
-   else
-     echo " "
-     echo "${str} did not pass the minimum criteria that passes ecoPCR reads to the next step."
-     rm ${str}
-   fi
-  done
-fi
+    str1=${str%_ecoPCR_blast_input_a_and_g_clean.fasta}
+    j=${str1#${ODIR}/${NAME}_ecoPCR/cleaned/}
+    mkdir -p ${ODIR}/${NAME}_ecoPCR/cleaned/${j}
 
-# need to check if the array is done before moving on to the next step
-if [ "${LOCALMODE}" = "FALSE" ];
-then
- filename="${ODIR}/${NAME}_BLAST/blast_complete_outfiles.txt"
- filelines=`cat $filename`
- echo Start
- for line in $filelines ; do
-  while ! [ -f ${line} ];
-  do
-   echo "files not ready"
-   sleep 600
+    if [ -s "${str}" ];
+    then
+      mkdir -p ${ODIR}/${NAME}_BLAST/
+      mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out
+      mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out/raw
+      mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out/fasta
+      echo " "
+      echo "${str} has ecoPCR reads that passed the minimum criteria to move to the next step."
+      # split ecopcr output into files with 500 reads each
+      split -l 1000 ${str} ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_
+      # do something as file has data
+      i=1
+      for nam in ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_*
+      do
+        nam1=${nam%_*}
+    # submit blast jobs for each file, and then remove reads with duplicate accession version numbers
+        cp ${nam} ${nam1}_${i}
+        rm ${nam}
+        echo "${nam1}_${i}_blast_1.complete" >> ${ODIR}/${NAME}_BLAST/blast_complete_outfiles.txt
+        echo "${nam1}_${i}_blast_2.complete" >> ${ODIR}/${NAME}_BLAST/blast_complete_outfiles.txt
+        ((i=i+1))
+      done
+      # if local blast in line
+      # if not local submit blast array jobs
+      ### count number of files in the directory
+      file_count=$( shopt -s nullglob ; set -- ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_* ; echo $#)
+      # submit blast jobs for each file, and then remove reads with duplicate accession version numbers
+      array_var="\$SGE_TASK_ID"
+      printf "${BLAST1_HEADER}\n#$ -t 1-${file_count}\n\n\n/bin/bash ${DB}/scripts/sub_blast1.sh -n ${NAME} -q ${nam1}_${array_var} -o ${ODIR} -k ${j} -l blast_ready_${array_var} -d ${DB} -v ${EVAL1:=$BLAST1_eVALUE} -t ${THREAD1:=$BLAST1_NUM_THREADS} -i ${ID1:=$BLAST1_PERC_IDENTITY} -c ${COV1:=$BLAST1_HSP_PERC} -a ${RETURN:=$BLAST1_NUM_ALIGNMENTS} -y ${GO:=$GAP_OPEN} -z ${GE:=$GAP_EXTEND}\n" > ${ODIR}/Run_info/blast_jobs/blast1_${j}.sh
+      ${QUEUESUBMIT} ${ODIR}/Run_info/blast_jobs/blast1_${j}.sh
+      printf "${BLAST2_HEADER}\n#$ -t 1-${file_count}\n\n\n/bin/bash ${DB}/scripts/sub_blast2.sh -n ${NAME} -q ${nam1}_${array_var} -o ${ODIR} -k ${j} -l blast_ready_${array_var} -d ${DB} -w ${EVAL2:=$BLAST2_eVALUE} -j ${THREAD2:=$BLAST2_NUM_THREADS} -p ${ID2:=$BLAST2_PERC_IDENTITY} -x ${COV2:=$BLAST2_HSP_PERC} -a ${RETURN:=$BLAST2_NUM_ALIGNMENTS} -y ${GO:=$GAP_OPEN} -z ${GE:=$GAP_EXTEND}\n" > ${ODIR}/Run_info/blast_jobs/blast2_${j}.sh
+      ${QUEUESUBMIT} ${ODIR}/Run_info/blast_jobs/blast2_${j}.sh
+    else
+      echo " "
+      echo "${str} did not pass the minimum criteria that passes ecoPCR reads to the next step."
+      rm ${str}
+    fi
   done
- done
+  else
+    for str in ${ODIR}/${NAME}_ecoPCR/cleaned/*_ecoPCR_blast_input_a_and_g_clean.fasta
+    do
+    str1=${str%_ecoPCR_blast_input_a_and_g_clean.fasta}
+    j=${str1#${ODIR}/${NAME}_ecoPCR/cleaned/}
+    mkdir -p ${ODIR}/${NAME}_ecoPCR/cleaned/${j}
+    echo ""
+    echo "run BLAST inline"
+    if [ -s "${str}" ];
+    then
+      mkdir -p ${ODIR}/${NAME}_BLAST/
+      mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out
+      mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out/raw
+      mkdir -p ${ODIR}/${NAME}_BLAST/${j}_BLAST_out/fasta
+      echo " "
+      echo "${str} has ecoPCR reads that passed the minimum criteria."
+      # split ecopcr output into files with 500 reads each
+      split -l 1000 ${str} ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_
+        # do something as file has data
+      i=1
+      for nam in ${ODIR}/${NAME}_ecoPCR/cleaned/${j}/blast_ready_*
+      do
+        nam1=${nam%_*}
+      # submit blast jobs for each file, and then remove reads with duplicate accession version numbers
+        cp ${nam} ${nam1}_${i}
+        #  rm ${nam}
+        echo ""
+        echo "Running BLAST1 on ${nam1}_${i}"
+        date
+        printf "#!/bin/bash\n\n\n/bin/bash ${DB}/scripts/sub_blast1.sh -n ${NAME} -q ${nam1}_${i} -o ${ODIR} -k ${j} -l blast_ready_${i} -d ${DB} -v ${EVAL1:=$BLAST1_eVALUE} -t ${THREAD1:=$BLAST1_NUM_THREADS} -i ${ID1:=$BLAST1_PERC_IDENTITY} -c ${COV1:=$BLAST1_HSP_PERC} -a ${RETURN:=$BLAST1_NUM_ALIGNMENTS} -y ${GO:=$GAP_OPEN} -z ${GE:=$GAP_EXTEND}\n" > ${ODIR}/Run_info/blast_jobs/blast1_${j}_${i}.sh
+        /bin/bash ${ODIR}/Run_info/blast_jobs/blast1_${j}_${i}.sh
+        echo "Running BLAST2 on ${nam1}_${i}"
+        date
+        printf "#!/bin/bash\n\n\n/bin/bash ${DB}/scripts/sub_blast2.sh -n ${NAME} -q ${nam1}_${i} -o ${ODIR} -k ${j} -l blast_ready_${i} -d ${DB} -w ${EVAL2:=$BLAST2_eVALUE} -j ${THREAD2:=$BLAST2_NUM_THREADS} -p ${ID2:=$BLAST2_PERC_IDENTITY} -x ${COV2:=$BLAST2_HSP_PERC} -a ${RETURN:=$BLAST2_NUM_ALIGNMENTS} -y ${GO:=$GAP_OPEN} -z ${GE:=$GAP_EXTEND}\n" > ${ODIR}/Run_info/blast_jobs/blast2_${j}_${i}.sh
+        /bin/bash ${ODIR}/Run_info/blast_jobs/blast2_${j}_${i}.sh
+        ((i=i+1))
+      done
+      date
+    else
+      echo " "
+      echo "${str} did not pass the minimum criteria that passes ecoPCR reads to the next step."
+      rm ${str}
+    fi
+    done
+  fi
+
+  # need to check if the array is done before moving on to the next step
+  if [ "${LOCALMODE}" = "FALSE" ];
+  then
+  filename="${ODIR}/${NAME}_BLAST/blast_complete_outfiles.txt"
+  filelines=`cat $filename`
+  echo Start
+  for line in $filelines ; do
+    while ! [ -f ${line} ];
+    do
+    echo "files not ready"
+    sleep 600
+    done
+  done
+  else
+    echo ""
+  fi
 else
-  echo ""
+  echo "Skipping BLAST steps..."
 fi
 
 ###
@@ -384,29 +379,38 @@ fi
 echo " "
 echo " "
 echo "Part 2.1: Cleaning up blast results"
-echo "For each set of BLAST 1 and 2 results"
-echo "     Merge and De-replicate by NCBI accession version numbers, and convert to fasta format."
-echo "     Then use entrez-qiime to generate a corresponding taxonomy file, and clean the blast output and taxonomy file to eliminate poorly annotated sequences."
 mkdir -p ${ODIR}/${NAME}_db_filtered/${NAME}_fasta_and_taxonomy/
 mkdir -p ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy
 
-# for each BLAST_out fasta folder
-for str in ${ODIR}/${NAME}_BLAST/${NAME}_*_out
-do
-  str1=${str%_BLAST_out}
-  j=${str1#${ODIR}/${NAME}_BLAST/${NAME}_}
-  echo "${str}"
-  echo "${j}"
-  # reduce size blast files by combining and dereplicating by length.
-  python ${DB}/scripts/combine_and_dereplicate_fasta.py -o ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_${j}_blast_out.fasta -a ${str}/fasta/*_out.fasta
-done
+if [ "${SKIPBLAST}" = "FALSE" ]
+then
+  echo "For each set of BLAST 1 and 2 results"
+  echo "     Merge and De-replicate by NCBI accession version numbers, and convert to fasta format."
+  echo "     Then use entrez-qiime to generate a corresponding taxonomy file, and clean the blast output and taxonomy file to eliminate poorly annotated sequences."
+  # for each BLAST_out fasta folder
+  for str in ${ODIR}/${NAME}_BLAST/${NAME}_*_out
+  do
+    str1=${str%_BLAST_out}
+    j=${str1#${ODIR}/${NAME}_BLAST/${NAME}_}
+    echo "${str}"
+    echo "${j}"
+    # reduce size blast files by combining and dereplicating by length.
+    python ${DB}/scripts/combine_and_dereplicate_fasta.py -o ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_${j}_blast_out.fasta -a ${str}/fasta/*_out.fasta
+  done
 
-# for all of of derepliated blast hits, combine and depreplicate by length
-python ${DB}/scripts/combine_and_dereplicate_fasta.py -o ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_.fasta -a ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/*_blast_out.fasta
-rm ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/*_blast_out.fasta
+  # for all of of derepliated blast hits, combine and depreplicate by length
+  python ${DB}/scripts/combine_and_dereplicate_fasta.py -o ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_.fasta -a ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/*_blast_out.fasta
+  rm ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/*_blast_out.fasta
+else
+  echo "Creating fasta format from blast_db"
+  # create reference fasta from blast
+  blastdbcmd -entry all -db ${BLAST_DB} -out ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_.fasta
+fi
 
 ### add taxonomy using entrez_qiime.py
 echo "...Running ${j} entrez-qiime and cleaning up fasta and taxonomy files"
+echo python ${ENTREZ_QIIME} -i ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_.fasta -o ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_taxonomy -n ${TAXO} -a ${A2T} -r superkingdom,phylum,class,order,family,genus,species
+
 python ${ENTREZ_QIIME} -i ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_.fasta -o ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_taxonomy -n ${TAXO} -a ${A2T} -r superkingdom,phylum,class,order,family,genus,species
 # clean up reads based on low resolution taxonomy and store filtered reads in filtered file
 python ${DB}/scripts/clean_blast.py ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_.fasta ${ODIR}/${NAME}_db_filtered/${NAME}_fasta_and_taxonomy/${NAME}_.fasta ${ODIR}/${NAME}_db_unfiltered/${NAME}_fasta_and_taxonomy/${NAME}_taxonomy.txt ${ODIR}/${NAME}_db_filtered/${NAME}_fasta_and_taxonomy/${NAME}_taxonomy.txt
